chore: refactor chart components
