+++ examples-standalone/grid-charts-integration/src/app/charts/day/day-chart.component.ts
@@ -1,0 +1,15 @@
+import {
+    Component,
+    Input
+} from "@angular/core";
+
+
+@Component({
+    selector: "day-chart",
+    templateUrl: './day-chart.template.html'
+})
+export class DayChartComponent {
+    @Input() public data: number[];
+    @Input() public changePct: number;
+
+}
